requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
transformers==4.30.0
torch==2.0.1
pydantic==1.10.7
pandas==1.5.3
python-dotenv==1.0.0
neo4j==5.7.0
streamlit==1.21.0
